Audio preview hot fix

+++ gst/audio_inference_elements/base/audio_infer_impl.h
@@ -18,7 +18,7 @@ class AudioInferImpl {
     void setNumOfSamplesToSlide();
 
   private:
-    std::vector<int16_t> audioData;
+    std::vector<float> audioData;
     std::vector<ulong> inferenceStartTime;
     bool startTimeSet = false;
     GvaAudioBaseInference *audio_base_inference;
